move duplicated incremental/full-table code to common
