+++ tap_oracle/sync_strategies/full_table.py
@@ -4,6 +4,7 @@ from singer import utils, write_message, get_bookmark
 import singer.metadata as metadata
 from singer.schema import Schema
 import tap_oracle.db as orc_db
+import tap_oracle.sync_strategies.common as common
 import singer.metrics as metrics
 import copy
 import pdb
@@ -15,41 +16,9 @@ LOGGER = singer.get_logger()
 
 UPDATE_BOOKMARK_PERIOD = 1000
 
-def row_to_singer_message(stream, row, version, columns, time_extracted):
-   row_to_persist = ()
-   for idx, elem in enumerate(row):
-      property_type = stream.schema.properties[columns[idx]].type
-      if elem is None:
-            row_to_persist += (elem,)
-      elif 'integer' in property_type or property_type == 'integer':
-         integer_representation = int(elem)
-         row_to_persist += (integer_representation,)
-      else:
-         row_to_persist += (elem,)
-
-   rec = dict(zip(columns, row_to_persist))
-
-   return singer.RecordMessage(
-        stream=stream.stream,
-        record=rec,
-        version=version,
-        time_extracted=time_extracted)
-
-def OutputTypeHandler(cursor, name, defaultType, size, precision, scale):
-    if defaultType == cx_Oracle.NUMBER:
-        return cursor.var(decimal.Decimal, arraysize = cursor.arraysize)
-
-
-def prepare_columns_sql(stream, c):
-   column_name = """ "{}" """.format(c)
-   if 'string' in stream.schema.properties[c].type and stream.schema.properties[c].format == 'date-time':
-      return "to_char({})".format(column_name)
-
-   return column_name
-
 def sync_view(conn_config, stream, state, desired_columns):
    connection = orc_db.open_connection(conn_config)
-   connection.outputtypehandler = OutputTypeHandler
+   connection.outputtypehandler = common.OutputTypeHandler
 
    cur = connection.cursor()
    cur.execute("ALTER SESSION SET TIME_ZONE = '00:00'")
@@ -73,7 +42,7 @@ def sync_view(conn_config, stream, state, desired_columns):
    md = metadata.to_map(stream.metadata)
    schema_name = md.get(()).get('schema-name')
 
-   escaped_columns = map(lambda c: prepare_columns_sql(stream, c), desired_columns)
+   escaped_columns = map(lambda c: common.prepare_columns_sql(stream, c), desired_columns)
    escaped_schema  = schema_name
    escaped_table   = stream.table
    activate_version_message = singer.ActivateVersionMessage(
@@ -90,7 +59,11 @@ def sync_view(conn_config, stream, state, desired_columns):
 
       LOGGER.info("select %s", select_sql)
       for row in cur.execute(select_sql):
-         record_message = row_to_singer_message(stream, row, nascent_stream_version, desired_columns, time_extracted)
+         record_message = common.row_to_singer_message(stream,
+                                                       row,
+                                                       nascent_stream_version,
+                                                       desired_columns,
+                                                       time_extracted)
          singer.write_message(record_message)
          counter.increment()
 
@@ -102,7 +75,7 @@ def sync_view(conn_config, stream, state, desired_columns):
 
 def sync_table(conn_config, stream, state, desired_columns):
    connection = orc_db.open_connection(conn_config)
-   connection.outputtypehandler = OutputTypeHandler
+   connection.outputtypehandler = common.OutputTypeHandler
 
    cur = connection.cursor()
    cur.execute("ALTER SESSION SET TIME_ZONE = '00:00'")
@@ -131,7 +104,7 @@ def sync_table(conn_config, stream, state, desired_columns):
    md = metadata.to_map(stream.metadata)
    schema_name = md.get(()).get('schema-name')
 
-   escaped_columns = map(lambda c: prepare_columns_sql(stream, c), desired_columns)
+   escaped_columns = map(lambda c: common.prepare_columns_sql(stream, c), desired_columns)
    escaped_schema  = schema_name
    escaped_table   = stream.table
    activate_version_message = singer.ActivateVersionMessage(
@@ -165,7 +138,11 @@ def sync_table(conn_config, stream, state, desired_columns):
       for row in cur.execute(select_sql):
          ora_rowscn = row[-1]
          row = row[:-1]
-         record_message = row_to_singer_message(stream, row, nascent_stream_version, desired_columns, time_extracted)
+         record_message = common.row_to_singer_message(stream,
+                                                       row,
+                                                       nascent_stream_version,
+                                                       desired_columns,
+                                                       time_extracted)
 
          singer.write_message(record_message)
          state = singer.write_bookmark(state, stream.tap_stream_id, 'ORA_ROWSCN', ora_rowscn)
